requirements.txt
torch>=2.0
torchvision
opencv-python
tqdm
matplotlib
scikit-learn
numpy
pillow
flask
matplotlib
